# backend/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: WebTui Chat Backend API
  version: 0.9.0
  description: Contract nền cho backend WebTui Chat. Tài liệu API được cập nhật theo từng phase backend.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /desktop/releases/{channel}/{target}/{arch}/{current_version}:
    get:
      security: []
      summary: Manifest signed updater cho desktop app
      description: Trả manifest Tauri updater đã publish theo channel/target/arch. Trả 204 khi chưa cấu hình hoặc chưa có bản mới.
      parameters:
        - name: channel
          in: path
          required: true
          schema:
            type: string
            enum: [stable, beta]
        - name: target
          in: path
          required: true
          schema:
            type: string
            example: windows-x86_64
        - name: arch
          in: path
          required: true
          schema:
            type: string
            example: x86_64
        - name: current_version
          in: path
          required: true
          schema:
            type: string
            example: 0.1.0
      responses:
        "200":
          description: Manifest có bản cập nhật signed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DesktopUpdaterManifest"
        "204":
          description: Không có bản cập nhật hoặc chưa cấu hình manifest
  /mobile/releases/{platform}/{channel}/{current_version}:
    get:
      security: []
      summary: Metadata phiên bản mobile app
      description: Trả manifest mobile nếu đã publish, hoặc policy fallback từ cấu hình backend.
      parameters:
        - name: platform
          in: path
          required: true
          schema:
            type: string
            enum: [android, ios]
        - name: channel
          in: path
          required: true
          schema:
            type: string
            enum: [stable, beta, internal]
        - name: current_version
          in: path
          required: true
          schema:
            type: string
            example: 1.0.0
      responses:
        "200":
          description: Metadata hoặc manifest phiên bản mobile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MobileReleaseManifest"
        "204":
          description: Chưa có manifest mobile phù hợp
  /downloads/manifest/{channel}:
    get:
      security: []
      summary: Manifest public cho file download
      description: Dùng cho APK/Desktop direct download để mobile/web kiểm tra checksum trước khi tải.
      parameters:
        - name: channel
          in: path
          required: true
          schema:
            type: string
            enum: [stable, beta, internal]
      responses:
        "200":
          description: Manifest download public
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadManifest"
        "204":
          description: Chưa cấu hình download manifest
...
  /api/v1/workspaces/{workspace_id}/sync:
    get:
      summary: Catch-up realtime event cho mobile
      description: Mobile gọi sau khi mở app, reconnect hoặc đổi workspace để lấy event mới sau cursor.
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - name: cursor
          in: query
          schema:
            type: string
            format: uuid
        - name: device_id
          in: query
          schema:
            type: string
          description: Nếu không gửi cursor, backend dùng cursor đã ack của device này.
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 200
      responses:
        "200":
          description: Danh sách event cần đồng bộ
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SyncCatchUp"
  /api/v1/workspaces/{workspace_id}/sync/ack:
    post:
      summary: Ack cursor đồng bộ của một thiết bị
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncAckRequest"
      responses:
        "200":
          description: Cursor đã được lưu
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SyncAck"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    WorkspaceID:
      name: workspace_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          description: Dữ liệu trả về
        error:
          $ref: "#/components/schemas/ApiError"
        meta:
          description: Metadata phân trang hoặc thông tin bổ sung
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    DesktopUpdaterManifest:
      type: object
      required: [version, platforms]
      properties:
        version:
          type: string
        notes:
          type: string
        pub_date:
          type: string
          format: date-time
        platforms:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/DesktopUpdaterPlatform"
    DesktopUpdaterPlatform:
      type: object
      required: [signature, url]
      properties:
        signature:
          type: string
        url:
          type: string
          format: uri
...
    MobileReleaseManifest:
      type: object
      properties:
        platform:
          type: string
          enum: [android, ios]
        channel:
          type: string
          enum: [stable, beta, internal]
        current_version:
          type: string
        minimum_version:
          type: string
        recommended_version:
          type: string
        version:
          type: string
        required:
          type: boolean
        download_url:
          type: string
          format: uri
        store_url:
          type: string
          format: uri
        checksum_sha256:
          type: string
        notes:
          type: string
    DownloadManifest:
      type: object
      properties:
        channel:
          type: string
        files:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              url:
                type: string
                format: uri
              byte_size:
                type: integer
              checksum_sha256:
                type: string
...
    SyncCatchUp:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/SyncEvent"
        next_cursor:
          type: string
          format: uuid
        has_more:
          type: boolean
        server_time:
          type: string
          format: date-time
    SyncEvent:
      type: object
      properties:
        event_id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        type:
          type: string
          example: MessageCreated
        aggregate_type:
          type: string
        aggregate_id:
          type: string
          format: uuid
        event_version:
          type: integer
        occurred_at:
          type: string
          format: date-time
        payload:
          type: object
          additionalProperties: true
    SyncAckRequest:
      type: object
      required: [device_id, cursor]
      properties:
        device_id:
          type: string
        cursor:
          type: string
          format: uuid
    SyncAck:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        device_id:
          type: string
        cursor_event_id:
          type: string
          format: uuid
        cursor_created_at:
          type: string
          format: date-time
        acked_at:
          type: string
          format: date-time
...